:art: 统一测试用例初始化方法
