+++ src/test/java/org/b3log/solo/processor/IndexProcessorTestCase.java
@@ -19,12 +19,9 @@ package org.b3log.solo.processor;
 
 import org.apache.commons.lang.StringUtils;
 import org.b3log.latke.Keys;
-import org.b3log.latke.model.User;
 import org.b3log.solo.AbstractTestCase;
 import org.b3log.solo.model.Option;
-import org.b3log.solo.service.InitService;
-import org.b3log.solo.service.UserQueryService;
-import org.json.JSONObject;
+import org.b3log.solo.service.StatisticMgmtService;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
@@ -54,17 +51,7 @@ public class IndexProcessorTestCase extends AbstractTestCase {
      */
     @Test
     public void init() throws Exception {
-        final InitService initService = getInitService();
-
-        final JSONObject requestJSONObject = new JSONObject();
-        requestJSONObject.put(User.USER_EMAIL, "test@gmail.com");
-        requestJSONObject.put(User.USER_NAME, "Admin");
-        requestJSONObject.put(User.USER_PASSWORD, "pass");
-
-        initService.init(requestJSONObject);
-
-        final UserQueryService userQueryService = getUserQueryService();
-        Assert.assertNotNull(userQueryService.getUserByEmailOrUserName("test@gmail.com"));
+        super.init();
     }
 
     /**
@@ -152,5 +139,35 @@ public class IndexProcessorTestCase extends AbstractTestCase {
 
         final String content = stringWriter.toString();
         Assert.assertTrue(StringUtils.contains(content, "<title>Admin 的个人博客</title>"));
+    }
+
+    /**
+     * showIndex.
+     *
+     * @throws Exception exception
+     */
+    @Test(dependsOnMethods = "init")
+    public void onlineVisitorCountRefresher() throws Exception {
+        final HttpServletRequest request = mock(HttpServletRequest.class);
+        when(request.getServletContext()).thenReturn(mock(ServletContext.class));
+        when(request.getRequestURI()).thenReturn("/console/stat/onlineVisitorRefresh");
+        when(request.getMethod()).thenReturn("GET");
+        when(request.getAttribute(Keys.TEMAPLTE_DIR_NAME)).thenReturn(Option.DefaultPreference.DEFAULT_SKIN_DIR_NAME);
+        when(request.getAttribute(Keys.HttpRequest.START_TIME_MILLIS)).thenReturn(System.currentTimeMillis());
+
+        final MockDispatcherServlet dispatcherServlet = new MockDispatcherServlet();
+        dispatcherServlet.init();
+
+        final StringWriter stringWriter = new StringWriter();
+        final PrintWriter printWriter = new PrintWriter(stringWriter);
+
+        final HttpServletResponse response = mock(HttpServletResponse.class);
+        when(response.getWriter()).thenReturn(printWriter);
+
+        StatisticMgmtService.ONLINE_VISITORS.put("127.0.0.1", Long.MIN_VALUE);
+
+        dispatcherServlet.service(request, response);
+
+        Assert.assertTrue(0 == StatisticMgmtService.ONLINE_VISITORS.size());
     }
 }
