+++ src/test/java/org/b3log/solo/processor/MockDispatcherServlet.java
@@ -17,23 +17,19 @@
  */
 package org.b3log.solo.processor;
 
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import org.b3log.latke.servlet.HTTPRequestContext;
 import org.b3log.latke.servlet.HttpControl;
-import org.b3log.latke.servlet.handler.AdviceHandler;
-import org.b3log.latke.servlet.handler.ArgsHandler;
-import org.b3log.latke.servlet.handler.Handler;
-import org.b3log.latke.servlet.handler.MethodInvokeHandler;
-import org.b3log.latke.servlet.handler.RequestDispatchHandler;
-import org.b3log.latke.servlet.handler.RequestPrepareHandler;
+import org.b3log.latke.servlet.handler.*;
 import org.b3log.latke.servlet.renderer.AbstractHTTPResponseRenderer;
 import org.b3log.latke.servlet.renderer.HTTP404Renderer;
 import org.b3log.latke.servlet.renderer.HTTP500Renderer;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * Mock dispatcher servlet for unit tests.
