--- src/test/java/org/b3log/solo/processor/StatProcessorTestCase.java
@@ -1,96 +1,0 @@
-/*
- * Solo - A small and beautiful blogging system written in Java.
- * Copyright (c) 2010-2018, b3log.org & hacpai.com
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-package org.b3log.solo.processor;
-
-import java.io.PrintWriter;
-import java.io.StringWriter;
-import javax.servlet.ServletContext;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import org.b3log.latke.Keys;
-import org.b3log.latke.model.User;
-import org.b3log.solo.AbstractTestCase;
-import org.b3log.solo.model.Option;
-import org.b3log.solo.service.InitService;
-import org.b3log.solo.service.StatisticMgmtService;
-import org.b3log.solo.service.UserQueryService;
-import org.json.JSONObject;
-import static org.mockito.Mockito.*;
-import org.testng.Assert;
-import org.testng.annotations.Test;
-
-/**
- * {@link StatProcessor} test case.
- *
- * @author <a href="http://88250.b3log.org">Liang Ding</a>
- * @version 1.0.1.0, Feb 18, 2017
- * @since 1.7.0
- */
-@Test(suiteName = "processor")
-public class StatProcessorTestCase extends AbstractTestCase {
-
-    /**
-     * Init.
-     *
-     * @throws Exception exception
-     */
-    @Test
-    public void init() throws Exception {
-        final InitService initService = getInitService();
-
-        final JSONObject requestJSONObject = new JSONObject();
-        requestJSONObject.put(User.USER_EMAIL, "test@gmail.com");
-        requestJSONObject.put(User.USER_NAME, "Admin");
-        requestJSONObject.put(User.USER_PASSWORD, "pass");
-
-        initService.init(requestJSONObject);
-
-        final UserQueryService userQueryService = getUserQueryService();
-        Assert.assertNotNull(userQueryService.getUserByEmailOrUserName("test@gmail.com"));
-    }
-
-    /**
-     * showIndex.
-     *
-     * @throws Exception exception
-     */
-    @Test(dependsOnMethods = "init")
-    public void showIndex() throws Exception {
-        final HttpServletRequest request = mock(HttpServletRequest.class);
-        when(request.getServletContext()).thenReturn(mock(ServletContext.class));
-        when(request.getRequestURI()).thenReturn("/console/stat/onlineVisitorRefresh");
-        when(request.getMethod()).thenReturn("GET");
-        when(request.getAttribute(Keys.TEMAPLTE_DIR_NAME)).thenReturn(Option.DefaultPreference.DEFAULT_SKIN_DIR_NAME);
-        when(request.getAttribute(Keys.HttpRequest.START_TIME_MILLIS)).thenReturn(System.currentTimeMillis());
-
-        final MockDispatcherServlet dispatcherServlet = new MockDispatcherServlet();
-        dispatcherServlet.init();
-
-        final StringWriter stringWriter = new StringWriter();
-        final PrintWriter printWriter = new PrintWriter(stringWriter);
-
-        final HttpServletResponse response = mock(HttpServletResponse.class);
-        when(response.getWriter()).thenReturn(printWriter);
-
-        StatisticMgmtService.ONLINE_VISITORS.put("127.0.0.1", Long.MIN_VALUE);
-        
-        dispatcherServlet.service(request, response);
-
-        Assert.assertTrue(0 == StatisticMgmtService.ONLINE_VISITORS.size());
-    }
-}
