+++ src/test/java/org/b3log/solo/repository/ArchiveDateRepositoryImplTestCase.java
@@ -22,7 +22,6 @@ import org.apache.commons.lang.time.DateUtils;
 import org.b3log.latke.repository.Transaction;
 import org.b3log.solo.AbstractTestCase;
 import org.b3log.solo.model.ArchiveDate;
-import org.b3log.solo.repository.ArchiveDateRepository;
 import org.json.JSONObject;
 import org.testng.annotations.Test;
