+++ src/test/java/org/b3log/solo/repository/ArticleRepositoryImplTestCase.java
@@ -22,7 +22,6 @@ import org.b3log.latke.repository.Query;
 import org.b3log.latke.repository.Transaction;
 import org.b3log.solo.AbstractTestCase;
 import org.b3log.solo.model.Article;
-import org.b3log.solo.repository.ArticleRepository;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import org.testng.Assert;
