+++ src/test/java/org/b3log/solo/repository/LinkRepositoryImplTestCase.java
@@ -20,7 +20,6 @@ package org.b3log.solo.repository;
 import org.b3log.latke.repository.Transaction;
 import org.b3log.solo.AbstractTestCase;
 import org.b3log.solo.model.Link;
-import org.b3log.solo.repository.LinkRepository;
 import org.json.JSONObject;
 import org.testng.Assert;
 import org.testng.annotations.Test;
@@ -36,7 +35,7 @@ public final class LinkRepositoryImplTestCase extends AbstractTestCase {
 
     /**
      * Tests.
-     * 
+     *
      * @throws Exception exception
      */
     @Test
