+++ src/test/java/org/b3log/solo/repository/PageRepositoryImplTestCase.java
@@ -21,7 +21,6 @@ import org.b3log.latke.Keys;
 import org.b3log.latke.repository.Transaction;
 import org.b3log.solo.AbstractTestCase;
 import org.b3log.solo.model.Page;
-import org.b3log.solo.repository.PageRepository;
 import org.json.JSONObject;
 import org.testng.Assert;
 import org.testng.annotations.Test;
