+++ src/test/java/org/b3log/solo/service/CommentMgmtServiceTestCase.java
@@ -17,9 +17,7 @@
  */
 package org.b3log.solo.service;
 
-import java.util.List;
 import org.b3log.latke.Keys;
-import org.b3log.latke.model.User;
 import org.b3log.latke.util.Requests;
 import org.b3log.solo.AbstractTestCase;
 import org.b3log.solo.model.Comment;
@@ -28,6 +26,8 @@ import org.json.JSONArray;
 import org.json.JSONObject;
 import org.testng.Assert;
 import org.testng.annotations.Test;
+
+import java.util.List;
 
 /**
  * {@link CommentMgmtService} test case.
@@ -40,27 +40,17 @@ public class CommentMgmtServiceTestCase extends AbstractTestCase {
 
     /**
      * Init.
-     * 
+     *
      * @throws Exception exception
      */
     @Test
     public void init() throws Exception {
-        final InitService initService = getInitService();
-
-        final JSONObject requestJSONObject = new JSONObject();
-        requestJSONObject.put(User.USER_EMAIL, "test@gmail.com");
-        requestJSONObject.put(User.USER_NAME, "Admin");
-        requestJSONObject.put(User.USER_PASSWORD, "pass");
-
-        initService.init(requestJSONObject);
-
-        final UserQueryService userQueryService = getUserQueryService();
-        Assert.assertNotNull(userQueryService.getUserByEmailOrUserName("test@gmail.com"));
+        super.init();
     }
 
     /**
      * Add Article Comment.
-     * 
+     *
      * @throws Exception exception
      */
     @Test(dependsOnMethods = "init")
@@ -103,7 +93,7 @@ public class CommentMgmtServiceTestCase extends AbstractTestCase {
 
     /**
      * Add Page Comment.
-     * 
+     *
      * @throws Exception exception
      */
     @Test(dependsOnMethods = "addArticleComment")
@@ -125,7 +115,7 @@ public class CommentMgmtServiceTestCase extends AbstractTestCase {
 
         Assert.assertNotNull(result);
         Assert.assertEquals(result.getJSONArray(Comment.COMMENTS).length(),
-                            2);  // 2 article comments
+                2);  // 2 article comments
 
         final CommentMgmtService commentMgmtService = getCommentMgmtService();
         final JSONObject requestJSONObject = new JSONObject();
@@ -148,7 +138,7 @@ public class CommentMgmtServiceTestCase extends AbstractTestCase {
 
         Assert.assertNotNull(result);
         Assert.assertEquals(result.getJSONArray(Comment.COMMENTS).length(),
-                            3);  // 2 article comments + 1 page comment
+                3);  // 2 article comments + 1 page comment
 
         final List<JSONObject> pageComments = commentQueryService.getComments(pageId);
         Assert.assertNotNull(pageComments);
@@ -157,7 +147,7 @@ public class CommentMgmtServiceTestCase extends AbstractTestCase {
 
     /**
      * Adds a page.
-     * 
+     *
      * @throws Exception exception
      */
     private void addPage() throws Exception {
