+++ src/test/java/org/b3log/solo/service/LinkMgmtServiceTestCase.java
@@ -17,7 +17,6 @@
  */
 package org.b3log.solo.service;
 
-import org.b3log.latke.model.User;
 import org.b3log.solo.AbstractTestCase;
 import org.b3log.solo.model.Link;
 import org.json.JSONObject;
@@ -35,27 +34,17 @@ public class LinkMgmtServiceTestCase extends AbstractTestCase {
 
     /**
      * Init.
-     * 
+     *
      * @throws Exception exception
      */
     @Test
     public void init() throws Exception {
-        final InitService initService = getInitService();
-
-        final JSONObject requestJSONObject = new JSONObject();
-        requestJSONObject.put(User.USER_EMAIL, "test@gmail.com");
-        requestJSONObject.put(User.USER_NAME, "Admin");
-        requestJSONObject.put(User.USER_PASSWORD, "pass");
-
-        initService.init(requestJSONObject);
-
-        final UserQueryService userQueryService = getUserQueryService();
-        Assert.assertNotNull(userQueryService.getUserByEmailOrUserName("test@gmail.com"));
+        super.init();
     }
 
     /**
      * Add Link.
-     * 
+     *
      * @throws Exception exception
      */
     @Test(dependsOnMethods = "init")
@@ -76,7 +65,7 @@ public class LinkMgmtServiceTestCase extends AbstractTestCase {
 
     /**
      * Remove Link.
-     * 
+     *
      * @throws Exception exception
      */
     @Test(dependsOnMethods = "init")
@@ -109,7 +98,7 @@ public class LinkMgmtServiceTestCase extends AbstractTestCase {
 
     /**
      * Update Link.
-     * 
+     *
      * @throws Exception exception
      */
     @Test(dependsOnMethods = "init")
@@ -145,7 +134,7 @@ public class LinkMgmtServiceTestCase extends AbstractTestCase {
 
     /**
      * Change Order.
-     * 
+     *
      * @throws Exception exception
      */
     @Test(dependsOnMethods = "addLink")
