+++ src/test/java/org/b3log/solo/service/LinkQueryServiceTestCase.java
@@ -17,7 +17,6 @@
  */
 package org.b3log.solo.service;
 
-import org.b3log.latke.model.User;
 import org.b3log.latke.util.Requests;
 import org.b3log.solo.AbstractTestCase;
 import org.b3log.solo.model.Link;
@@ -36,27 +35,17 @@ public class LinkQueryServiceTestCase extends AbstractTestCase {
 
     /**
      * Init.
-     * 
+     *
      * @throws Exception exception
      */
     @Test
     public void init() throws Exception {
-        final InitService initService = getInitService();
-
-        final JSONObject requestJSONObject = new JSONObject();
-        requestJSONObject.put(User.USER_EMAIL, "test@gmail.com");
-        requestJSONObject.put(User.USER_NAME, "Admin");
-        requestJSONObject.put(User.USER_PASSWORD, "pass");
-
-        initService.init(requestJSONObject);
-
-        final UserQueryService userQueryService = getUserQueryService();
-        Assert.assertNotNull(userQueryService.getUserByEmailOrUserName("test@gmail.com"));
+        super.init();
     }
 
     /**
      * Add Link.
-     * 
+     *
      * @throws Exception exception
      */
     @Test(dependsOnMethods = "init")
@@ -77,12 +66,12 @@ public class LinkQueryServiceTestCase extends AbstractTestCase {
 
     /**
      * Get Links.
-     * 
+     *
      * @throws Exception exception
      */
     @Test(dependsOnMethods = "addLink")
     public void getLinks() throws Exception {
-       final LinkQueryService linkQueryService = getLinkQueryService();
+        final LinkQueryService linkQueryService = getLinkQueryService();
 
         final JSONObject paginationRequest =
                 Requests.buildPaginationRequest("1/10/20");
