+++ src/test/java/org/b3log/solo/service/TagMgmtServiceTestCase.java
@@ -17,10 +17,7 @@
  */
 package org.b3log.solo.service;
 
-import org.b3log.latke.model.User;
 import org.b3log.solo.AbstractTestCase;
-import org.json.JSONObject;
-import org.testng.Assert;
 import org.testng.annotations.Test;
 
 /**
@@ -34,27 +31,17 @@ public class TagMgmtServiceTestCase extends AbstractTestCase {
 
     /**
      * Init.
-     * 
+     *
      * @throws Exception exception
      */
     @Test
     public void init() throws Exception {
-        final InitService initService = getInitService();
-
-        final JSONObject requestJSONObject = new JSONObject();
-        requestJSONObject.put(User.USER_EMAIL, "test@gmail.com");
-        requestJSONObject.put(User.USER_NAME, "Admin");
-        requestJSONObject.put(User.USER_PASSWORD, "pass");
-
-        initService.init(requestJSONObject);
-
-        final UserQueryService userQueryService = getUserQueryService();
-        Assert.assertNotNull(userQueryService.getUserByEmailOrUserName("test@gmail.com"));
+        super.init();
     }
 
     /**
      * Remove Unused Tags.
-     * 
+     *
      * @throws Exception exception
      */
     @Test(dependsOnMethods = "init")
