+++ src/test/java/org/b3log/solo/service/TagQueryServiceTestCase.java
@@ -17,14 +17,13 @@
  */
 package org.b3log.solo.service;
 
-import java.util.List;
-
-import org.b3log.latke.model.User;
 import org.b3log.solo.AbstractTestCase;
 import org.b3log.solo.model.Tag;
 import org.json.JSONObject;
 import org.testng.Assert;
 import org.testng.annotations.Test;
+
+import java.util.List;
 
 /**
  * {@link TagQueryService} test case.
@@ -37,27 +36,17 @@ public class TagQueryServiceTestCase extends AbstractTestCase {
 
     /**
      * Init.
-     * 
+     *
      * @throws Exception exception
      */
     @Test
     public void init() throws Exception {
-        final InitService initService = getInitService();
-
-        final JSONObject requestJSONObject = new JSONObject();
-        requestJSONObject.put(User.USER_EMAIL, "test@gmail.com");
-        requestJSONObject.put(User.USER_NAME, "Admin");
-        requestJSONObject.put(User.USER_PASSWORD, "pass");
-
-        initService.init(requestJSONObject);
-
-        final UserQueryService userQueryService = getUserQueryService();
-        Assert.assertNotNull(userQueryService.getUserByEmailOrUserName("test@gmail.com"));
+        super.init();
     }
 
     /**
      * Get Tags.
-     * 
+     *
      * @throws Exception exception
      */
     @Test(dependsOnMethods = "init")
@@ -72,19 +61,19 @@ public class TagQueryServiceTestCase extends AbstractTestCase {
 
     /**
      * Get Tag By Title.
-     * 
+     *
      * @throws Exception exception
      */
     @Test(dependsOnMethods = "init")
     public void getTagByTitle() throws Exception {
         final TagQueryService tagQueryService = getTagQueryService();
-        
+
         final JSONObject result = tagQueryService.getTagByTitle("Solo");
         Assert.assertNotNull(result);
-        
+
         final JSONObject tag = result.getJSONObject(Tag.TAG);
         Assert.assertNotNull(tag);
         Assert.assertEquals(tag.getString(Tag.TAG_TITLE), "Solo");
-        
+
     }
 }
